refactor namespaces - ServerGridEditor -> Atlas.ServerGridEditor and AtlasGridDataLibrary -> Atlas.GridData
