+++ Src/ServerGridEditor/Code/DiscoveryZone.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Drawing;
-using AtlasGridDataLibrary;
+using Atlas.GridData;
 
-namespace ServerGridEditor
+namespace Atlas.ServerGridEditor
 {
     public static class DiscoveryZoneInstanceEx
     {
