+++ Src/ServerGridEditor/Code/GlobalSettings.cs
@@ -1,6 +1,6 @@
 ﻿using System.IO;
 
-namespace ServerGridEditor
+namespace Atlas.ServerGridEditor
 {
     public class GlobalSettings
     {
