+++ Src/ServerGridEditor/Forms/EditAllLocksForm.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Windows.Forms;
 
-namespace ServerGridEditor.Forms
+namespace Atlas.ServerGridEditor.Forms
 {
     public partial class EditAllLocksForm : Form
     {
