+++ Src/ServerGridEditor/Forms/EditIslandInstance.cs
@@ -1,9 +1,9 @@
-﻿using AtlasGridDataLibrary;
+﻿using Atlas.GridData;
 using System;
 using System.Collections.Generic;
 using System.Windows.Forms;
 
-namespace ServerGridEditor.Forms
+namespace Atlas.ServerGridEditor.Forms
 {
     public partial class EditIslandInstance : Form
     {
