+++ Src/ServerGridEditor/Forms/EditServerTemplate.cs
@@ -1,10 +1,10 @@
-﻿using AtlasGridDataLibrary;
+﻿using Atlas.GridData;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Windows.Forms;
 
-namespace ServerGridEditor
+namespace Atlas.ServerGridEditor
 {
     public partial class EditServerTemplate : Form
     {
