+++ Src/ServerGridEditor/Forms/EditServerTemplates.cs
@@ -1,9 +1,9 @@
-﻿using AtlasGridDataLibrary;
+﻿using Atlas.GridData;
 using System;
 using System.Drawing;
 using System.Windows.Forms;
 
-namespace ServerGridEditor.Forms
+namespace Atlas.ServerGridEditor.Forms
 {
     public partial class EditServerTemplates : Form
     {
